[Chore]: 폴더&파일 이름 변경 및 ridingView 생성
