+++ Tourding_FE/Resources/Assets.xcassets/Home/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
