+++ Tourding_FE/Resources/Assets.xcassets/Home/startpoint.imageset/Contents.json
